🚚 refactor apollo client location
